Add files via upload
That definitely doesn't look right. One troublesome practical issue is that when the ray is in the material with the higher refractive index, there is no real solution to Snell’s law, and thus there is no refraction possible. If we refer back to Snell's law and the derivation of sin𝜃′
sin
⁡
θ
′
:

sin𝜃′=𝜂𝜂′⋅sin𝜃
sin
⁡
θ
′
=
η
η
′
⋅
sin
⁡
θ

If the ray is inside glass and outside is air (𝜂=1.5
η
=
1.5
 and 𝜂′=1.0
η
′
=
1.0
):

sin𝜃′=1.51.0⋅sin𝜃
sin
⁡
θ
′
=
1.5
1.0
⋅
sin
⁡
θ

The value of sin𝜃′
sin
⁡
θ
′
 cannot be greater than 1. So, if,

1.51.0⋅sin𝜃>1.0
1.5
1.0
⋅
sin
⁡
θ
>
1.0
,

the equality between the two sides of the equation is broken, and a solution cannot exist. If a solution does not exist, the glass cannot refract, and therefore must reflect the ray:

if (refraction_ratio * sin_theta > 1.0) {
    // Must Reflect
    ...
} else {
    // Can Refract
    ...
}

Here all the light is reflected, and because in practice that is usually inside solid objects, it is called “total internal reflection”. This is why sometimes the water-air boundary acts as a perfect mirror when you are submerged.

We can solve for sin_theta using the trigonometric qualities:
sin𝜃=1−cos2𝜃‾‾‾‾‾‾‾‾‾‾√
sin
⁡
θ
=
1
−
cos
2
⁡
θ

and

cos𝜃=𝐑⋅𝐧
cos
⁡
θ
=
R
⋅
n

double cos_theta = fmin(dot(-unit_direction, rec.normal), 1.0);
double sin_theta = sqrt(1.0 - cos_theta*cos_theta);

if (refraction_ratio * sin_theta > 1.0) {
    // Must Reflect
    ...
} else {
    // Can Refract
    ...
}
Listing 57: [material.h] Determining if the ray can refract